# PERF: fix some of .clip() performance regression by using numpy arrays where possible