@jreback 2019-01-19
@qwhelan can you merge master and ping when passing